+++ public/index.php
@@ -27,5 +27,7 @@ require __DIR__ . '/../src/app/vulnerabilities/Magento Downloader.php';
 require __DIR__ . '/../src/app/vulnerabilities/Swagger Config File.php';
 require __DIR__ . '/../src/app/vulnerabilities/WP Debug.php';
 require __DIR__ . '/../src/app/vulnerabilities/AWStats Script.php';
+require __DIR__ . '/../src/app/vulnerabilities/API Key Scanner.php';
+require __DIR__ . '/../src/app/vulnerabilities/Database Connection String.php';
 
 $app->run();