+++ CHANGELOG.md
@@ -27,6 +27,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Vulnerability: Swagger Config File
 - Vulnerability: Wordpress Debugger
 - Vulnerability: AWStats Script
+- Vulnerability: API Key Scanner
+- Vulnerability: Database Connection String
 - GIF Favicon
 - Unauthenticated Gitlab SSRF CVE 2021-22214 Demonstration
 - Software Versions List